Add an icon for .xlaunch documents
It seems we need this

When we register Xlaunch directly as the handler for .xlaunch files,
Windows kindly makes a document icon for us, consisting of the xlaunch
program icon on top of a blank page.

However, if we register "run bash -l -c xlaunch" as the handler for
.xlaunch files, Windows makes the document icon from the run icon, and
there doesn't seem to any way to suggest it use a different icon.

So, provide an explicit .xlaunch document icon in the executable, and
then we can register that as the icon for .xlaunch files at the same time
as we register the handler for .xlaunch  files.